moved release_notes.py out of gen.py
